Série OK

+++ pc/src/serie/SerialInterface.java
@@ -15,6 +15,6 @@ public interface SerialInterface
 	public void communiquer(byte[] out);
 	public void close();
 	public boolean available() throws IOException;
-	public byte read() throws IOException, MissingCharacterException;
+	public int read() throws IOException, MissingCharacterException;
 	public int getFirstID();
 }
